+++ src/main/java/repository/DBConnection.java
@@ -1,4 +1,4 @@
-package db;
+package repository;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
