+++ src/main/java/repository/JdbcBoardRepository.java
@@ -1,4 +1,4 @@
-package db;
+package repository;
 
 import domain.Position;
 import domain.Team;
